Refactor call list cancel endpoint and update schema
Refactored the cancel call list endpoint to remove the path variable and accept competitionId in the request body. Updated the CallListService to return an Either result. Modified the database schema to add 'cancelled' as a call_type, created new tables for function and match_day, and removed the session_referees table. Updated test data and schema scripts accordingly.

reports are commented becuase tests callList are in the first stage but have reports for testing purposes